+++ firmware/src/src/ble.cpp
@@ -24,7 +24,7 @@
 
 #include "io.h"
 #include "log.h"
-#include "nano33ble.h"
+#include "htmain.h"
 #include "soc_flash.h"
 #include "trackersettings.h"
