+++ firmware/src/src/btjoystick.cpp
@@ -28,7 +28,7 @@
 #include "io.h"
 #include "joystick.h"
 #include "log.h"
-#include "nano33ble.h"
+#include "htmain.h"
 #include "trackersettings.h"
