+++ firmware/src/src/btparahead.cpp
@@ -27,7 +27,7 @@
 #include "defines.h"
 #include "io.h"
 #include "log.h"
-#include "nano33ble.h"
+#include "htmain.h"
 #include "blechars.h"
 #include "trackersettings.h"
