+++ firmware/src/src/btpararmt.cpp
@@ -27,7 +27,7 @@
 
 #include "io.h"
 #include "log.h"
-#include "nano33ble.h"
+#include "htmain.h"
 #include "opentxbt.h"
 #include "trackersettings.h"
