+++ firmware/src/src/sense.cpp
@@ -32,7 +32,7 @@
 #include "io.h"
 #include "joystick.h"
 #include "log.h"
-#include "nano33ble.h"
+#include "htmain.h"
 #include "pmw.h"
 #include "soc_flash.h"
 #include "trackersettings.h"
